Add reporting in csv format at the end of the run
This can be checked by running:

```
./scripts/setup.sh mint.json owner.json
sleep 15 # wait for the transactions to be finalized
./scripts/audit.sh mint.json owner.json
```

The summary mode makes a few assumptions about how the fraud would look,
which may not always be correct.  At worst, it should flag some false
positives if there are multiple owner reassignments for one account.

The detail view is meant to show all of the important transactions
found, so it's always possible to double-check what the summary mode
produced.